refactor(benchmark): split execute.py into execute/profiling/accuracy/pipeline
execute.py concentrated all CLI benchmark orchestration; the large
_run_benchmark_async and _score_accuracy dominated its complexity. Split along
natural seams into four cohesive modules:

- profiling.py (new): profile-trigger protocol (vLLM /start_profile,/stop_profile)
  + ProfileController (URL derivation + start/stop/payload lifecycle)
- accuracy.py (new): AccuracyConfiguration, _score_accuracy, _load_osl_backend,
  write_accuracy_results
- pipeline.py (new): MetricsPipeline — ZMQ + metrics-aggregator/event-logger
  subprocess lifecycle (start/drain_and_build_report/abort/close) + snapshot→Report
- execute.py: thin orchestrator

Everything audit.py/cli.py import stays in execute (TestMode, BenchmarkResult,
_salvage_tmpfs, setup/run/finalize, run_benchmark, resolve_report_dir); one-way
import graph (execute → {profiling, accuracy, pipeline}). Three unit test files
repointed.

Rebased onto main after the SWE-bench scorer (#342). #342's execute.py accuracy
changes are routed into their post-split homes: AccuracyConfiguration
model_params/endpoint_config fields, _effective_external_sample_count,
_accuracy_uuid_bound (None-on-unavailable), and SKIP_ENDPOINT_PHASE handling in
_score_accuracy land in accuracy.py; _validate_accuracy_config_for_scorer, the
external-scorer skips/logging, and the dataset-loader/preflight wiring land in
execute.py.

Also restores the early-stopping aggregator flag the split had dropped:
"--early-stopping" is now threaded through pipeline._build_aggregator_args (gated
on settings.early_stopping.enabled), replacing an orphaned append to a variable
that no longer existed in _run_benchmark_async after the split.

Verified: unit suite green except pre-existing macOS-only failures
(sched_setaffinity / IPC-socket-path in cpu_affinity/token_metrics/metrics_aggregator,
none in commands/benchmark). ruff/ruff-format/prettier/license clean; mypy clean on
the changed files (remaining mypy errors are the macOS-only os.sched_* attr-defined
false-positives in untouched files).
