+++ src/inference_endpoint/commands/benchmark/execute.py
@@ -19,6 +19,10 @@ Phases:
     1. setup_benchmark()        — load tokenizer, dataset, config (no IO)
     2. run_benchmark_async()    — HTTP client + async BenchmarkSession
     3. finalize_benchmark()     — accuracy scoring, results JSON
+
+Cohesive sub-concerns live in sibling modules: profiler triggers (``profiling``),
+accuracy scoring (``accuracy``), and the ZMQ/metrics/event-logger service lifecycle
+(``pipeline``).
 """
 
 from __future__ import annotations
@@ -26,59 +30,45 @@ from __future__ import annotations
 import asyncio
 import json
 import logging
-import numbers
 import random
 import shutil
 import signal
 import tempfile
-import time
 import uuid
-from collections.abc import Callable, Iterable
+from collections.abc import Callable
 from dataclasses import dataclass, field
 from dataclasses import replace as dataclass_replace
 from datetime import datetime
 from pathlib import Path
-from typing import Any, TextIO
-from urllib import error as urllib_error
-from urllib import request as urllib_request
+from typing import Any
 from urllib.parse import urljoin
 
-import msgspec
 import msgspec.json
 import msgspec.structs
 from huggingface_hub import model_info
 from tqdm import tqdm
 from transformers.utils import logging as transformers_logging
 
-from inference_endpoint.async_utils.event_publisher import EventPublisherService
 from inference_endpoint.async_utils.loop_manager import LoopManager
-from inference_endpoint.async_utils.services.launcher import (
-    ServiceConfig,
-    ServiceLauncher,
+from inference_endpoint.commands.benchmark.accuracy import (
+    AccuracyConfiguration,
+    _effective_external_sample_count,
+    _score_accuracy,
+    write_accuracy_results,
 )
-from inference_endpoint.async_utils.services.metrics_aggregator.snapshot import (
-    snapshot_to_dict,
+from inference_endpoint.commands.benchmark.pipeline import MetricsPipeline
+from inference_endpoint.commands.benchmark.profiling import (
+    ProfileController,
+    _write_profiling_section,
 )
-from inference_endpoint.async_utils.services.metrics_aggregator.subscriber import (
-    MetricsSnapshotSubscriber,
-)
-from inference_endpoint.async_utils.services.metrics_aggregator.token_metrics import (
-    encode_lengths,
-    load_reference_backend,
-)
-from inference_endpoint.async_utils.transport.zmq.context import ManagedZMQContext
 from inference_endpoint.compliance import AuditRunSpec
 from inference_endpoint.config.runtime_settings import RuntimeSettings
 from inference_endpoint.config.schema import (
     APIType,
     BenchmarkConfig,
     DatasetType,
-    EndpointConfig,
     LoadPattern,
     LoadPatternType,
-    ModelParams,
-    ProfilerEngine,
-    ScorerMethod,
     StreamingMode,
     TestMode,
     TestType,
@@ -93,7 +83,6 @@ from inference_endpoint.endpoint_client.cpu_affinity import AffinityPlan, pin_lo
 from inference_endpoint.endpoint_client.http_client import HTTPEndpointClient
 from inference_endpoint.endpoint_client.http_sample_issuer import HttpClientSampleIssuer
 from inference_endpoint.evaluation import Extractor
-from inference_endpoint.evaluation.accuracy_results import average_accuracy
 from inference_endpoint.evaluation.scoring import Scorer
 from inference_endpoint.exceptions import (
     ExecutionError,
@@ -110,8 +99,7 @@ from inference_endpoint.load_generator.session import (
     PhaseType,
     SessionResult,
 )
-from inference_endpoint.metrics.report import Report, series_metric_dict
-from inference_endpoint.utils.atomic_write import atomic_write_bytes
+from inference_endpoint.metrics.report import Report
 
 transformers_logging.set_verbosity_error()
 
@@ -171,34 +159,6 @@ class BenchmarkResult:
     # settings.profiling.engine is set; None otherwise. Rendered into
     # report.txt and a sibling profiling.json by finalize_benchmark.
     profiling: dict[str, Any] | None = None
-
-
-@dataclass
-class AccuracyConfiguration:
-    scorer: type[Scorer]
-    extractor: type[Extractor] | None
-    dataset_name: str
-    dataset: Dataset
-    report_dir: Path
-    ground_truth_column: str | None
-    num_repeats: int
-    extras: dict[str, Any] = field(default_factory=dict)
-    model_params: ModelParams | None = None
-    endpoint_config: EndpointConfig | None = None
-    # Discriminates the inline perf-scored entry (PERFORMANCE) from real accuracy
-    # datasets (ACCURACY). Branch on this, not on dataset_name == "performance":
-    # a dataset legitimately named "performance" must not be misclassified.
-    dataset_type: DatasetType = DatasetType.ACCURACY
-
-
-def _effective_external_sample_count(
-    eval_cfg: AccuracyConfiguration,
-) -> int | None:
-    """Clamp an external scorer's requested count to its loaded dataset size."""
-    count = eval_cfg.scorer.external_sample_count(eval_cfg.extras)
-    if count is None:
-        return None
-    return min(count, eval_cfg.dataset.num_samples())
 
 
 @dataclass
@@ -378,6 +338,9 @@ def _load_datasets(
             num_repeats=acc_cfg.accuracy_config.num_repeats,
             **loader_kwargs,
         )
+        # Value/api-type validity of the override is already enforced at config
+        # construction (BenchmarkConfig validates each dataset's effective params),
+        # so this cannot raise for a validated config.
         ds_model_params = acc_cfg.effective_generation_config(config.model_params)
         ds.load(api_type=config.endpoint_config.api_type, model_params=ds_model_params)
         logger.info(f"Loaded {ds} - {ds.num_samples()} samples")
@@ -705,27 +668,6 @@ def _build_phases(
     return phases
 
 
-def _load_final_snapshot_from_disk(path: Path) -> dict[str, Any] | None:
-    """Read the persisted ``final_snapshot.json`` written by the aggregator.
-
-    Returns the snapshot in its dict form — the same shape produced by
-    ``snapshot_to_dict`` and consumed by ``Report.from_snapshot``. No
-    intermediate Struct decode (see ``Report.from_snapshot`` docstring
-    for why the dict shape is the consumer contract).
-
-    Returns ``None`` if the file is missing (the aggregator was killed
-    by an uncatchable signal — SIGKILL, OOM-kill — before its handler
-    could write) or unreadable.
-    """
-    if not path.exists():
-        return None
-    try:
-        return json.loads(path.read_bytes())
-    except Exception as e:  # noqa: BLE001 — best-effort.
-        logger.warning("Failed to read final snapshot %s: %s", path, e)
-        return None
-
-
 class _PerfPhaseTimeout:
     """Session-stop timer that bounds the PERFORMANCE phase only.
 
@@ -760,108 +702,90 @@ class _PerfPhaseTimeout:
             self._handle = None
 
 
-# (start_path, stop_path) for each supported inference engine's profiling
-# protocol. Add a row when introducing a new ProfilerEngine variant.
-_PROFILE_PATHS: dict[ProfilerEngine, tuple[str, str]] = {
-    ProfilerEngine.VLLM: ("/start_profile", "/stop_profile"),
-}
-
-
-def _derive_profile_urls(
-    endpoints: list[str], engine: ProfilerEngine, action: str
-) -> list[str]:
-    """One profile URL per endpoint, derived from the engine's HTTP protocol.
-
-    For vLLM: strip a trailing ``/v1`` from each endpoint and append
-    ``/{start,stop}_profile``. ``action`` is ``"start"`` or ``"stop"``.
-    """
-    if not endpoints:
-        raise ValueError(
-            f"profiling.engine={engine.value} but endpoint_config.endpoints "
-            f"is empty; cannot derive {action} URLs"
-        )
-    start_path, stop_path = _PROFILE_PATHS[engine]
-    path = start_path if action == "start" else stop_path
-    urls: list[str] = []
-    for ep in endpoints:
-        base = ep.rstrip("/")
-        if base.endswith("/v1"):
-            base = base[:-3]
-        urls.append(f"{base.rstrip('/')}{path}")
-    return urls
-
-
-def _post_profile(url: str) -> dict[str, Any]:
-    """POST {url} with empty body; never raises. Returns a record dict suitable
-    for report.txt rendering and profiling.json serialization."""
-    record: dict[str, Any] = {
-        "url": url,
-        "sent_at_ns": time.monotonic_ns(),
-        "sent_at_iso": datetime.now().isoformat(timespec="milliseconds"),
-        "status": None,
-        "error": None,
-    }
-    req = urllib_request.Request(url, method="POST", data=b"")
+async def _create_issuer(
+    ctx: BenchmarkContext, loop: asyncio.AbstractEventLoop
+) -> tuple[HttpClientSampleIssuer, HTTPEndpointClient]:
+    """Create the HTTP endpoint client + sample issuer, or raise SetupError."""
+    config = ctx.config
+    endpoints = config.endpoint_config.endpoints
+    logger.info(f"Connecting: {endpoints}")
     try:
-        with urllib_request.urlopen(req, timeout=2) as resp:
-            record["status"] = resp.status
-    except urllib_error.HTTPError as e:
-        record["status"] = e.code
-        record["error"] = f"{e.code} {e.reason}"
-    except Exception as e:  # noqa: BLE001 — profile failures must never abort a run
-        record["error"] = f"{type(e).__name__}: {e}"
-    return record
-
-
-def _render_profile_status(rec: dict[str, Any]) -> str:
-    status = rec.get("status")
-    error = rec.get("error")
-    if status == 200:
-        return "200 OK"
-    if status == 404:
-        return (
-            "404 (profiling not enabled on server — pass "
-            "--profiler-config.profiler=... to server)"
-        )
-    if error:
-        return error
-    if status is not None:
-        return str(status)
-    return "ERROR"
-
-
-def _write_profiling_section(f: TextIO, profiling: dict[str, Any]) -> None:
-    """Append the Profiling section to report.txt (called after report.display)."""
-    starts = profiling.get("starts", [])
-    stops = profiling.get("stops", [])
-    f.write("\n------------------- Profiling -------------------\n")
-    f.write(f"Engine: {profiling.get('engine', 'unknown')}\n")
-    f.write("Start:\n")
-    for rec in starts:
-        f.write(
-            f"  POST {rec['url']} @ {rec['sent_at_iso']} → "
-            f"{_render_profile_status(rec)}\n"
-        )
-    if stops:
-        f.write("Stop:\n")
-        for rec in stops:
-            suffix = (
-                " (from abort handler)" if rec.get("stop_reason") == "abort" else ""
+        api_type: APIType = config.endpoint_config.api_type
+        # client.api_type is propagated from endpoint_config.api_type by
+        # BenchmarkConfig._propagate_client_api_type — no override needed here.
+        client_overrides: dict = {
+            "endpoint_urls": [
+                urljoin(e.rstrip("/") + "/", api_type.default_route())
+                for e in endpoints
+            ],
+            "api_key": config.endpoint_config.api_key,
+            "event_logs_dir": ctx.report_dir,
+            "cpu_affinity": ctx.affinity_plan,
+        }
+        if ctx.accuracy_only:
+            # Single-stream (num_workers=1, max_connections=1) is baked into
+            # config in setup_benchmark so it is persisted to config.yaml;
+            # no runtime override needed here.
+            logger.info(
+                "Accuracy-only: single-stream (1 worker, 1 connection) for "
+                "deterministic ordering"
             )
-            f.write(
-                f"  POST {rec['url']} @ {rec['sent_at_iso']} → "
-                f"{_render_profile_status(rec)}{suffix}\n"
-            )
-    if starts and stops:
-        first_start = min(r["sent_at_ns"] for r in starts)
-        last_stop = max(r["sent_at_ns"] for r in stops)
-        f.write(f"Trigger span: {(last_stop - first_start) / 1e9:.2f} s\n")
-    f.write(
-        "\nNote: actual trace window is bounded by server-side "
-        "--profiler-config.delay_iterations and "
-        "--profiler-config.max_iterations.\n"
-        "Trace artifact path is in server stdout.\n"
+        http_config = config.settings.client.with_updates(**client_overrides)
+        http_client = await HTTPEndpointClient.create(http_config, loop)
+        issuer = HttpClientSampleIssuer(http_client)
+        return issuer, http_client
+    except Exception as e:
+        raise SetupError(f"Failed to connect to endpoint: {e}") from e
+
+
+def _build_agentic_strategy(
+    ctx: BenchmarkContext,
+) -> AgenticInferenceStrategy | None:
+    """Build the agentic inference strategy when the perf dataset uses it."""
+    if not isinstance(ctx.dataloader, AgenticInferenceDataset):
+        return None
+    agentic_cfg = None
+    if ctx.config.datasets:
+        perf_ds_cfg = next(
+            (d for d in ctx.config.datasets if d.type == DatasetType.PERFORMANCE),
+            None,
+        )
+        if perf_ds_cfg is not None:
+            agentic_cfg = perf_ds_cfg.agentic_inference
+    assert ctx.dataloader.conversation_metadata is not None
+    return AgenticInferenceStrategy(
+        conversation_manager=ConversationManager(),
+        dataset_metadata=ctx.dataloader.conversation_metadata,
+        agentic_inference_config=agentic_cfg,
+        target_concurrency=ctx.config.settings.load_pattern.target_concurrency,
     )
+
+
+def _wire_on_sample_complete(
+    collector: ResponseCollector,
+    agentic_inference_strategy: AgenticInferenceStrategy | None,
+    publisher: Any,
+) -> Callable[[QueryResult], None]:
+    """Compose the per-sample completion callback (agentic strategy + collector)."""
+    if agentic_inference_strategy is None:
+        return collector.on_complete_hook
+
+    def _on_sample_complete(result: QueryResult) -> None:
+        try:
+            agentic_inference_strategy.on_sample_complete(result)
+        except Exception:
+            logger.exception(
+                "agentic_inference_strategy.on_sample_complete failed (result=%s)",
+                result.id,
+            )
+        try:
+            collector.on_complete_hook(result)
+        except Exception:
+            logger.exception("collector.on_complete_hook failed (result=%s)", result.id)
+
+    agentic_inference_strategy._session_on_sample_complete = _on_sample_complete
+    agentic_inference_strategy._session_publisher = publisher
+    return _on_sample_complete
 
 
 async def _run_benchmark_async(
@@ -880,218 +804,71 @@ async def _run_benchmark_async(
     )
     collector = ResponseCollector(collect_responses=ctx.collect_responses, pbar=pbar)
 
-    # ZMQ context for event publishing + service launcher
-    tmpfs_dir: Path | None = None
+    # Tmpfs for high-frequency writes (event log); execute owns its lifecycle
+    # (salvage + rmtree). metrics_output_dir lives on disk under the report dir so
+    # the final snapshot is preserved with the rest of the run artifacts — it is
+    # NOT tmpfs and is never removed here. Paths are computed here (no mkdir) so the
+    # cleanup `except` can always reference tmpfs_dir; the mkdirs run inside the try
+    # so a mkdir failure that already created tmpfs_dir is still salvaged/removed.
+    shm = Path("/dev/shm")
+    tmpfs_base = shm if shm.exists() else Path(tempfile.gettempdir())
+    tmpfs_dir = tmpfs_base / f"benchmark_{session_id}"
+    event_log_dir = tmpfs_dir / "events"
+    metrics_output_dir = ctx.report_dir / "metrics"
+
+    pipe = MetricsPipeline(
+        config,
+        tokenizer_name=ctx.tokenizer_name,
+        enable_streaming=ctx.enable_streaming,
+        event_log_dir=event_log_dir,
+        metrics_output_dir=metrics_output_dir,
+        loop=loop,
+    )
+    report: Report | None = None
+    profiler: ProfileController
+
     try:
-        with ManagedZMQContext.scoped(io_threads=2) as zmq_ctx:
-            # Event publisher
-            publisher = EventPublisherService(zmq_ctx)
-            pub_socket_name = publisher.socket_name
-
-            # Tmpfs for high-frequency writes (event log).
-            shm = Path("/dev/shm")
-            use_shm = shm.exists()
-            tmpfs_base = shm if use_shm else Path(tempfile.gettempdir())
-            tmpfs_dir = tmpfs_base / f"benchmark_{session_id}"
+        try:
             tmpfs_dir.mkdir(parents=True, exist_ok=True)
-
-            event_log_dir = tmpfs_dir / "events"
             event_log_dir.mkdir(parents=True, exist_ok=True)
-
-            # Metrics-snapshot output (disk fallback for the final snapshot).
-            # Lives under the report dir so it's preserved with the rest of
-            # the run artifacts.
-            metrics_output_dir = ctx.report_dir / "metrics"
             metrics_output_dir.mkdir(parents=True, exist_ok=True)
-
-            metrics_socket_name = f"metrics_pub_{uuid.uuid4().hex[:8]}"
-
-            # Connect the metrics-snapshot subscriber BEFORE launching the
-            # aggregator subprocess that binds the matching PUB socket. ZMQ
-            # tolerates connect-before-bind on IPC (the connect resolves once
-            # the binder appears), and starting the SUB reader early gives
-            # the subscription handshake time to complete during the
-            # ~1-2 second subprocess-launch window. This eliminates the
-            # slow-joiner risk of dropping early live ticks (or the worst
-            # case: missing COMPLETE if the SUB handshake never warms up).
-            if zmq_ctx.socket_dir is None:
-                raise RuntimeError("ZMQ socket_dir must be set after publisher bind")
-            metrics_subscriber = MetricsSnapshotSubscriber(
-                metrics_socket_name, zmq_ctx, loop
+            # Pre-derive profile URLs (inside the run scope so a bad config —
+            # engine set but no endpoints — fails before the run yet still triggers
+            # the tmpfs/pbar cleanup in the except/finally below).
+            profiler = ProfileController(
+                config.settings.profiling.engine,
+                config.endpoint_config.endpoints,
+                config.settings.profiling.urls,
             )
-            metrics_subscriber.start()
+            await pipe.start()
+            # start() guarantees the publisher exists; narrow it for the type checker.
+            publisher = pipe.publisher
+            assert publisher is not None
 
-            # Launch service subprocesses
-            launcher = ServiceLauncher(zmq_ctx)
-            aggregator_args: list[str] = [
-                "--socket-dir",
-                zmq_ctx.socket_dir,
-                "--socket-name",
-                pub_socket_name,
-                "--metrics-socket",
-                metrics_socket_name,
-                "--metrics-output-dir",
-                str(metrics_output_dir),
-            ]
-            if ctx.enable_streaming:
-                aggregator_args.append("--streaming")
-            if ctx.tokenizer_name is not None:
-                aggregator_args.extend(["--tokenizer", ctx.tokenizer_name])
-            aggregator_args.extend(
-                ["--drain-timeout", str(config.settings.drain.metrics_drain_timeout_s)]
-            )
-            aggregator_args.extend(
-                [
-                    "--tokenizer-workers",
-                    str(config.settings.drain.metrics_tokenizer_workers),
-                ]
-            )
-            if config.settings.early_stopping.enabled:
-                aggregator_args.append("--early-stopping")
-
-            # EventLoggerService writes events.jsonl to tmpfs (high-frequency writes)
-            event_logger_args: list[str] = [
-                "--log-dir",
-                str(event_log_dir),
-                "--socket-dir",
-                zmq_ctx.socket_dir,
-                "--socket-name",
-                pub_socket_name,
-                "--writers",
-                "jsonl",
-            ]
-
-            await launcher.launch(
-                [
-                    ServiceConfig(
-                        module="inference_endpoint.async_utils.services.metrics_aggregator",
-                        args=aggregator_args,
-                    ),
-                    ServiceConfig(
-                        module="inference_endpoint.async_utils.services.event_logger",
-                        args=event_logger_args,
-                    ),
-                ],
-                timeout=config.settings.service_ready_timeout_s,
-            )
-
-            # Create endpoint client on the shared loop
-            endpoints = config.endpoint_config.endpoints
-            logger.info(f"Connecting: {endpoints}")
-            http_client: HTTPEndpointClient | None = None
             try:
-                api_type: APIType = config.endpoint_config.api_type
-                # client.api_type is propagated from endpoint_config.api_type by
-                # BenchmarkConfig._propagate_client_api_type — no override needed here.
-                client_overrides: dict = {
-                    "endpoint_urls": [
-                        urljoin(e.rstrip("/") + "/", api_type.default_route())
-                        for e in endpoints
-                    ],
-                    "api_key": config.endpoint_config.api_key,
-                    "event_logs_dir": ctx.report_dir,
-                    "cpu_affinity": ctx.affinity_plan,
-                }
-                if ctx.accuracy_only:
-                    # Single-stream (num_workers=1, max_connections=1) is baked into
-                    # config in setup_benchmark so it is persisted to config.yaml;
-                    # no runtime override needed here.
-                    logger.info(
-                        "Accuracy-only: single-stream (1 worker, 1 connection) for "
-                        "deterministic ordering"
-                    )
-                http_config = config.settings.client.with_updates(**client_overrides)
-                http_client = await HTTPEndpointClient.create(http_config, loop)
-                issuer = HttpClientSampleIssuer(http_client)
-            except Exception as e:
-                pbar.close()
-                publisher.close()
-                launcher.kill_all()
-                raise SetupError(f"Failed to connect to endpoint: {e}") from e
+                issuer, http_client = await _create_issuer(ctx, loop)
+            except SetupError:
+                await pipe.abort()
+                raise
 
-            # Build agentic inference strategy if the performance dataset uses it.
-            agentic_inference_strategy: AgenticInferenceStrategy | None = None
-            if isinstance(ctx.dataloader, AgenticInferenceDataset):
-                agentic_cfg = None
-                if ctx.config.datasets:
-                    perf_ds_cfg = next(
-                        (
-                            d
-                            for d in ctx.config.datasets
-                            if d.type == DatasetType.PERFORMANCE
-                        ),
-                        None,
-                    )
-                    if perf_ds_cfg is not None:
-                        agentic_cfg = perf_ds_cfg.agentic_inference
-                assert ctx.dataloader.conversation_metadata is not None
-                agentic_inference_strategy = AgenticInferenceStrategy(
-                    conversation_manager=ConversationManager(),
-                    dataset_metadata=ctx.dataloader.conversation_metadata,
-                    agentic_inference_config=agentic_cfg,
-                    target_concurrency=ctx.config.settings.load_pattern.target_concurrency,
-                )
+            agentic_inference_strategy = _build_agentic_strategy(ctx)
+            on_sample_complete = _wire_on_sample_complete(
+                collector, agentic_inference_strategy, publisher
+            )
 
-            _on_sample_complete: Callable[[QueryResult], None]
-            if agentic_inference_strategy is not None:
-
-                def _on_sample_complete(result: QueryResult) -> None:
-                    try:
-                        agentic_inference_strategy.on_sample_complete(result)
-                    except Exception:
-                        logger.exception(
-                            "agentic_inference_strategy.on_sample_complete failed (result=%s)",
-                            result.id,
-                        )
-                    try:
-                        collector.on_complete_hook(result)
-                    except Exception:
-                        logger.exception(
-                            "collector.on_complete_hook failed (result=%s)", result.id
-                        )
-
-                agentic_inference_strategy._session_on_sample_complete = (
-                    _on_sample_complete
-                )
-                agentic_inference_strategy._session_publisher = publisher
-
-            else:
-                _on_sample_complete = collector.on_complete_hook
-
-            # Create session
             session = BenchmarkSession(
                 issuer=issuer,
                 event_publisher=publisher,
                 loop=loop,
-                on_sample_complete=_on_sample_complete,
+                on_sample_complete=on_sample_complete,
                 session_id=session_id,
             )
-
             phases = _build_phases(ctx, perf_strategy=agentic_inference_strategy)
-            report: Report | None = None
 
-            _timeout_done = False
             max_duration_ms = (
                 ctx.rt_settings.max_duration_ms if ctx.rt_settings is not None else None
             )
-
-            # Profile trigger state. Pre-derive URLs once so a bad config
-            # (engine set but no endpoints) fails before the run.
-            profiling_cfg = config.settings.profiling
-            profile_start_urls: list[str] = []
-            profile_stop_urls: list[str] = []
-            profile_starts: list[dict[str, Any]] = []
-            profile_stops: list[dict[str, Any]] = []
-            if profiling_cfg.engine is not None:
-                profile_endpoints = (
-                    profiling_cfg.urls or config.endpoint_config.endpoints
-                )
-                profile_start_urls = _derive_profile_urls(
-                    profile_endpoints, profiling_cfg.engine, "start"
-                )
-                profile_stop_urls = _derive_profile_urls(
-                    profile_endpoints, profiling_cfg.engine, "stop"
-                )
+            _timeout_done = False
             session_completed_normally = False
 
             def _on_global_timeout() -> None:
@@ -1115,20 +892,8 @@ async def _run_benchmark_async(
                 if phase.phase_type != PhaseType.PERFORMANCE:
                     return
                 # Fire /start_profile sequentially before any perf request is
-                # issued, so the server is armed when traffic begins. Blocks
-                # the loop briefly (sub-100ms per URL); strategy task hasn't
-                # been created yet so nothing is starved.
-                for url in profile_start_urls:
-                    rec = _post_profile(url)
-                    if rec["status"] == 200:
-                        logger.info("Profile start: %s -> 200 OK", url)
-                    else:
-                        logger.warning(
-                            "Profile start: %s -> %s",
-                            url,
-                            rec["error"] or rec["status"],
-                        )
-                    profile_starts.append(rec)
+                # issued, so the server is armed when traffic begins.
+                profiler.start()
 
             loop.add_signal_handler(signal.SIGINT, session.stop)
             try:
@@ -1141,134 +906,40 @@ async def _run_benchmark_async(
                 perf_timeout.cancel()
                 loop.remove_signal_handler(signal.SIGINT)
                 # Fire /stop_profile for URLs whose /start_profile succeeded.
-                # Unifies the clean phase-end path and the abort path —
-                # both reach this block, both fire stops.
-                if profile_starts:
-                    stop_reason = "phase_end" if session_completed_normally else "abort"
-                    for i, start_rec in enumerate(profile_starts):
-                        if start_rec["status"] != 200 or i >= len(profile_stop_urls):
-                            continue
-                        rec = _post_profile(profile_stop_urls[i])
-                        rec["stop_reason"] = stop_reason
-                        if rec["status"] == 200:
-                            logger.info(
-                                "Profile stop: %s -> 200 OK", profile_stop_urls[i]
-                            )
-                        else:
-                            logger.warning(
-                                "Profile stop: %s -> %s",
-                                profile_stop_urls[i],
-                                rec["error"] or rec["status"],
-                            )
-                        profile_stops.append(rec)
+                # Unifies the clean phase-end path and the abort path — both reach
+                # this block.
+                profiler.stop(session_completed_normally)
                 logger.info("Cleaning up...")
                 try:
                     if http_client:
                         await http_client.shutdown_async()
                 except Exception as e:
                     logger.warning(f"Client cleanup error: {e}")
-                logger.info(
-                    "Closing publisher (buffer=%d, pending=%d)...",
-                    publisher.buffered_count,
-                    publisher.pending_count,
-                )
-                publisher.close()
-                logger.info("Waiting for services to finish processing...")
-                await asyncio.to_thread(launcher.wait_for_exit, None)
-
-                # Source the snapshot dict for Report:
-                # 1. Preferred: the JSON file the aggregator atomically wrote
-                #    in publish_final (ENDED-driven or signal-handler-driven).
-                # 2. Fallback: convert the last live snapshot from pub/sub to
-                #    its dict form. Only reached when the aggregator was killed
-                #    by an uncatchable signal (SIGKILL / OOM) before its
-                #    handler could write. Report will be marked incomplete
-                #    because state will be LIVE / DRAINING, not "complete".
-                snap_dict: dict[str, Any] | None = _load_final_snapshot_from_disk(
-                    metrics_output_dir / "final_snapshot.json"
-                )
-                if snap_dict is not None:
-                    logger.info("Built report from final_snapshot.json")
-                elif metrics_subscriber.latest is not None:
-                    snap_dict = snapshot_to_dict(metrics_subscriber.latest)
-                    logger.warning(
-                        "No final_snapshot.json on disk; falling back to last "
-                        "pub/sub snapshot (state may or may not be terminal)"
-                    )
-                else:
-                    logger.error("No metrics snapshot available; cannot build report")
-
-                if snap_dict is not None:
-                    try:
-                        load_pattern = ctx.config.settings.load_pattern
-                        runtime_cfg = ctx.config.settings.runtime
-                        # load_pattern + warmup config and the RNG seeds, so
-                        # result_summary.json is self-describing and a valid run is
-                        # identified by its settings. The full, re-runnable config
-                        # lives in config.yaml alongside. The resolved/effective
-                        # runtime settings (sample count + ordering, which can differ
-                        # per audit phase) are deferred to a follow-up. endpoint_config
-                        # (api_key/URLs) is a sibling of settings and never included,
-                        # so no secrets.
-                        run_config = ctx.config.settings.model_dump(
-                            mode="json", include={"load_pattern", "warmup"}
-                        )
-                        run_config["scheduler_random_seed"] = (
-                            runtime_cfg.scheduler_random_seed
-                        )
-                        run_config["dataloader_random_seed"] = (
-                            runtime_cfg.dataloader_random_seed
-                        )
-                        report = Report.from_snapshot(
-                            snap_dict,
-                            run_config=run_config,
-                            use_legacy_loadgen_qps_metrics=(
-                                load_pattern.type == LoadPatternType.POISSON
-                                and load_pattern.use_legacy_loadgen_qps_metrics
-                            ),
-                        )
-                        if not report.complete:
-                            logger.warning(
-                                "Report is incomplete (state=%s, n_pending_tasks=%d)",
-                                report.state,
-                                snap_dict.get("n_pending_tasks", 0),
-                            )
-                        if report.legacy_loadgen_window_duration_ns is not None:
-                            logger.warning(
-                                "Reporting QPS/TPS with the legacy MLPerf LoadGen Server "
-                                "'completed' definition (deprecated; to be removed once a "
-                                "formal tail-cutting mechanism lands). Pass "
-                                "--no-use-legacy-loadgen-qps-metrics for endpoints-native "
-                                "metrics."
-                            )
-                    except Exception as e:  # noqa: BLE001 — best-effort report build.
-                        logger.warning(f"Failed to build report from snapshot: {e}")
-
-                metrics_subscriber.close()
+                # Graceful drain runs on both the clean-finish and session-failure
+                # paths (BenchmarkSession.run publishes ENDED in its own finally, so
+                # a failed run still has a terminal snapshot worth draining).
+                report = await pipe.drain_and_build_report()
+        finally:
+            # Cleanup runs on every path. pipe.close() (ZMQ scope exit) must run
+            # even if pbar.close() raises, and a failure here propagates to the
+            # outer except → tmpfs salvage — matching the monolith's guarantees
+            # (pbar.close was inside the ZMQ `with`, whose __exit__ always ran).
+            try:
                 pbar.close()
+            finally:
+                pipe.close()
     except BaseException:
-        # tmpfs_dir may still be None if the exception hit before it was
-        # created (e.g. ZMQ context setup), in which case there is nothing
-        # to clean up.
-        if tmpfs_dir is not None and tmpfs_dir.exists():
+        if tmpfs_dir.exists():
             _salvage_tmpfs(ctx.report_dir, tmpfs_dir)
             shutil.rmtree(tmpfs_dir, ignore_errors=True)
         raise
-
-    profiling_payload: dict[str, Any] | None = None
-    if profiling_cfg.engine is not None:
-        profiling_payload = {
-            "engine": profiling_cfg.engine.value,
-            "starts": profile_starts,
-            "stops": profile_stops,
-        }
 
     return BenchmarkResult(
         session=result,
         collector=collector,
         report=report,
         tmpfs_dir=tmpfs_dir,
-        profiling=profiling_payload,
+        profiling=profiler.payload(),
     )
 
 
@@ -1324,343 +995,35 @@ def _salvage_tmpfs(report_dir: Path, tmpfs_dir: Path) -> None:
         logger.debug(f"Copied {src_events} -> {dst_events}")
 
 
-def _phase_osl_stats(
-    sample_uuids: Iterable[str],
-    uuid_to_text: dict[str, str],
-    backend: Any,
-    batch_size: int = 256,
-) -> dict[str, Any] | None:
-    """Output-token-length rollup over one accuracy phase's completions.
+def _write_report_artifacts(
+    ctx: BenchmarkContext, report: Report, profiling: dict[str, Any] | None
+) -> None:
+    """Display the report and write result_summary.json + report.txt.
 
-    Counts tokens on each sample's response text via the shared reference
-    tokenizer backend — the server's ``completion_tokens`` is not persisted, only
-    the text is (in ``events.jsonl``) — then shapes the lengths via
-    ``series_metric_dict`` so the block matches the perf report's
-    ``output_sequence_lengths`` exactly. Returns ``None`` when the phase has no
-    completed outputs.
-
-    ``batch_size`` bounds each ``encode_batch`` pass: accuracy outputs can be tens
-    of thousands of tokens each (e.g. gpt-oss lcb at 32768), so counting the whole
-    population in one call would hold every Encoding in memory at once.
+    result_summary.json is the self-complete machine-readable report (carries
+    qps/tps/seeds/accuracy via Report.to_json); report.txt is the full
+    human-readable dump; the console log shows the summary.
     """
-    # Skip empty/failed completions (a failed request still logs a COMPLETE
-    # event with output == ""). The perf-side OslTrigger does the same
-    # (metrics_table.OslTrigger._extract_text returns None for empty text), so
-    # accuracy OSL matches its population and a failure isn't counted as a
-    # 0-token sample that would drag min/avg down.
-    texts = [
-        uuid_to_text[u] for u in sample_uuids if u in uuid_to_text and uuid_to_text[u]
-    ]
-    if not texts:
-        return None
-    lengths: list[int] = []
-    for i in range(0, len(texts), batch_size):
-        lengths.extend(encode_lengths(backend, texts[i : i + batch_size]))
-    return series_metric_dict(lengths) or None
+    report.display(fn=lambda s: logger.info(s), summary_only=True)
+    performance_dir = ctx.report_dir / "performance"
+    performance_dir.mkdir(parents=True, exist_ok=True)
+    report.to_json(save_to=performance_dir / "result_summary.json")
+
+    report_txt = ctx.report_dir / "report.txt"
+    with report_txt.open("w") as f:
+        report.display(fn=lambda s: print(s, file=f))
+        if profiling is not None:
+            _write_profiling_section(f, profiling)
+    logger.info("Report written to %s", report_txt)
 
 
-def _phase_response_counts(
-    sample_uuids: Iterable[str],
-    uuid_to_text: dict[str, str],
-) -> dict[str, int]:
-    """Per-phase response accounting over one accuracy phase's issued samples.
-
-    Complements :func:`_phase_osl_stats`, which reports token lengths only over
-    non-empty completions — on its own that can hide a run where the server
-    returned blanks or dropped requests. Classifies each issued ``sample_uuid``
-    as ``scored`` (COMPLETE, non-empty output — exactly the OSL population),
-    ``empty`` (COMPLETE with blank output: a failed request the load generator
-    logged as ERROR then an empty COMPLETE), or ``missing`` (no COMPLETE event).
-    ``issued == scored + empty + missing`` always holds.
-
-    Emptiness uses the same truthiness test as ``_phase_osl_stats`` so ``scored``
-    is byte-for-byte the OSL population — the two blocks cannot disagree.
-    """
-    issued = scored = empty = missing = 0
-    for u in sample_uuids:
-        issued += 1
-        if u not in uuid_to_text:
-            missing += 1
-        elif uuid_to_text[u]:
-            scored += 1
-        else:
-            empty += 1
-    return {"issued": issued, "scored": scored, "empty": empty, "missing": missing}
-
-
-def _accuracy_uuid_bound(
-    report_dir: Path | None, eval_configs: list[AccuracyConfiguration]
-) -> set[str] | None:
-    """Return issued accuracy UUIDs, or ``None`` when the map is unavailable.
-
-    An empty set is a valid bound for scorers that skip endpoint issuance.
-    """
-    if report_dir is None:
-        return None
-    try:
-        idx_map = msgspec.json.decode((report_dir / "sample_idx_map.json").read_bytes())
-    except (OSError, msgspec.DecodeError) as e:
-        logger.warning(
-            "Accuracy OSL uuid bound unavailable (%s); reading outputs unbounded", e
-        )
-        return None
-    # Invalid maps fall back to an unbounded read instead of failing scoring.
-    if not isinstance(idx_map, dict):
-        logger.warning(
-            "Accuracy OSL uuid bound: sample_idx_map.json is not an object; "
-            "reading outputs unbounded"
-        )
-        return None
-    bound: set[str] = set()
-    for ec in eval_configs:
-        if ec.dataset_type == DatasetType.ACCURACY:
-            per_dataset = idx_map.get(ec.dataset_name)
-            if isinstance(per_dataset, dict):
-                bound |= set(per_dataset)
-    return bound
-
-
-def _score_accuracy(
-    ctx: BenchmarkContext, result: SessionResult
-) -> list[dict[str, Any]]:
-    """Run configured scorers and return reportable accuracy entries.
-
-    One entry per eval_config, in order; no cross-dataset consolidation. Each
-    entry carries the scalar ``score`` plus sample accounting
-    (``unit_samples`` × ``num_repeats`` = ``total_samples``); a scorer that
-    returns a ``score_breakdown()`` (DeepSeek-R1, BFCL) also attaches
-    ``breakdown``. The ``"performance"`` inline entry totals the perf phases'
-    issued counts instead of unit × repeats (repeats is forced to 1 there).
-    """
-    accuracy_scores: list[dict[str, Any]] = []
-    eval_configs = [
-        eval_cfg
-        for eval_cfg in ctx.eval_configs
-        if not (ctx.test_mode == TestMode.PERF and eval_cfg.scorer.SKIP_ENDPOINT_PHASE)
-    ]
-
-    # Per-phase wall-clock (seconds) keyed by phase name. The accuracy phase name
-    # is the dataset_name; the inline-scored perf entry keys on "performance".
-    phase_durations: dict[str, float] = {}
-    for pr in result.phase_results:
-        phase_durations[pr.name] = phase_durations.get(pr.name, 0.0) + max(
-            0.0, (pr.end_time_ns - pr.start_time_ns) / 1e9
-        )
-
-    # Accuracy-phase output-token lengths (finalize-side, off the hot path): the
-    # aggregator only tokenizes perf-window samples, so count the accuracy
-    # responses (already in events.jsonl) here, using the same reference tokenizer
-    # as the perf side. (Counts still differ from perf for tool-call responses —
-    # client-side OSL is approximate for structured output.) Loaded only when a
-    # real accuracy dataset exists; a load failure or a tokenizer with no fast
-    # backend disables OSL rather than failing scoring.
-    has_accuracy = any(ec.dataset_type == DatasetType.ACCURACY for ec in eval_configs)
-    osl_backend: Any = None
-    if has_accuracy and ctx.tokenizer_name is not None:
-        try:
-            osl_backend = load_reference_backend(ctx.tokenizer_name)
-        except Exception as e:  # noqa: BLE001 - OSL is optional; never fail scoring
-            logger.warning(
-                "Accuracy OSL disabled: could not load tokenizer %r: %s",
-                ctx.tokenizer_name,
-                e,
-            )
-        else:
-            # A tokenizer with no fast (Rust) backend disables OSL rather than
-            # falling back to a slow Python-tokenizer count: the perf side
-            # (token_metrics._setup_shards) requires a fast backend too and raises
-            # without one, so OSL stays fast-only and consistent on both sides.
-            # Warn so the skip is visible instead of silently dropping the block.
-            if osl_backend is None:
-                logger.warning(
-                    "Accuracy OSL disabled: tokenizer %r has no fast (Rust) backend "
-                    "(token counting requires one, as on the perf side)",
-                    ctx.tokenizer_name,
-                )
-    # Bound the raw-output read to the accuracy population so finalize never holds
-    # the whole run's (incl. perf) response-text corpus.
-    accuracy_uuids = (
-        _accuracy_uuid_bound(ctx.report_dir, eval_configs) if has_accuracy else set()
-    )
-    uuid_to_text: dict[str, str] | None = None
-
-    for eval_cfg in eval_configs:
-        try:
-            scorer_kwargs = dict(eval_cfg.extras)
-            if (
-                getattr(eval_cfg.scorer, "SCORER_ID", None)
-                == ScorerMethod.SWE_BENCH.value
-            ):
-                scorer_kwargs.update(
-                    model_params=eval_cfg.model_params,
-                    endpoint_config=eval_cfg.endpoint_config,
-                )
-            scorer_instance = eval_cfg.scorer(
-                eval_cfg.dataset_name,
-                eval_cfg.dataset,
-                eval_cfg.report_dir,
-                extractor=eval_cfg.extractor,
-                ground_truth_column=eval_cfg.ground_truth_column,
-                **scorer_kwargs,
-            )
-        except TypeError as e:
-            raise InputValidationError(
-                f"Dataset '{eval_cfg.dataset_name}': invalid accuracy_config.extras "
-                f"for scorer '{eval_cfg.scorer.__name__}': {e}"
-            ) from e
-        score, n_repeats = scorer_instance.score()
-        # Coerce a numpy scalar score (np.float32/64, numpy ints — e.g. np.mean
-        # from the base Scorer) to a native Python float so the entry stays
-        # serializable by both msgspec (result_summary.json) and json
-        # (accuracy_results.json). numbers.Real catches every numpy scalar (not
-        # just np.float64, which isinstance(..., float) alone would miss) while
-        # leaving None / dict (RougeScorer) untouched; bool is excluded.
-        if isinstance(score, numbers.Real) and not isinstance(score, bool):
-            score = float(score)
-        unit_samples = eval_cfg.dataset.num_samples()
-        num_repeats = eval_cfg.num_repeats
-        if eval_cfg.dataset_type == DatasetType.PERFORMANCE:
-            # A performance dataset always scores its already-issued outputs once
-            # (enforced by the num_repeats == 1 guard in _load_datasets), so make
-            # that locally provable rather than relying on eval_cfg carrying 1.
-            num_repeats = 1
-            total_samples = sum(phase.issued_count for phase in result.perf_results)
-        else:
-            total_samples = unit_samples * num_repeats
-        if eval_cfg.scorer.SKIP_ENDPOINT_PHASE:
-            ext = _effective_external_sample_count(eval_cfg)
-            if ext is not None:
-                unit_samples = ext
-                total_samples = ext
-        entry: dict[str, Any] = {
-            "dataset_name": eval_cfg.dataset_name,
-            "extractor": (
-                eval_cfg.extractor.__name__ if eval_cfg.extractor is not None else None
-            ),
-            "ground_truth_column": eval_cfg.ground_truth_column,
-            "score": score,
-            "unit_samples": unit_samples,
-            "num_repeats": num_repeats,
-            "total_samples": total_samples,
-            # Wall-clock of this dataset's issue phase (seconds); 0.0 if the
-            # phase left no timing (e.g. a scored-but-not-issued dataset).
-            "duration_s": round(phase_durations.get(eval_cfg.dataset_name, 0.0), 3),
-            # False when the scorer produced only a partial headline (e.g.
-            # LegacyMLPerfDeepSeekR1Scorer when lcb-service was unreachable), so a
-            # partial number is never mistaken for a complete one.
-            "complete": scorer_instance.complete,
-            # Persist the same DatasetType discriminator carried on the eval config
-            # so consumers filter the inline perf-scored entry by type, not by
-            # matching dataset_name == "performance".
-            "dataset_type": eval_cfg.dataset_type.value,
-        }
-        breakdown = scorer_instance.score_breakdown()
-        if breakdown is not None:
-            entry["breakdown"] = breakdown
-
-        # Response accounting + avg/min/max output-token length. Skipped for the
-        # perf entry (its OSL / failure counts live in result_summary.json). The
-        # counts are computed independent of the tokenizer and of OSL returning a
-        # block — an all-failed phase must still publish scored=0 rather than
-        # silently omitting everything. OSL stays tokenizer-gated. A read/tokenize
-        # failure only drops these blocks — it never fails scoring.
-        if eval_cfg.dataset_type == DatasetType.ACCURACY:
-            try:
-                if uuid_to_text is None:
-                    # Built once from the first scorer and reused for every
-                    # dataset. get_raw_outputs() returns the model's actual
-                    # completion text (not the scorer's scoring-normalized form)
-                    # for *all* phases' COMPLETE events, bounded to the accuracy
-                    # population; intersecting it with each dataset's
-                    # sample_index_map yields correct per-dataset counts.
-                    out_df = scorer_instance.get_raw_outputs(accuracy_uuids)
-                    uuid_to_text = dict(
-                        zip(out_df["sample_uuid"], out_df["output"], strict=False)
-                    )
-                    # Drop the DataFrame so finalize doesn't hold both it and the
-                    # dict (each carrying the response-text corpus).
-                    del out_df
-                entry["response_counts"] = _phase_response_counts(
-                    scorer_instance.sample_index_map, uuid_to_text
-                )
-                if osl_backend is not None:
-                    t0 = time.perf_counter()
-                    osl = _phase_osl_stats(
-                        scorer_instance.sample_index_map, uuid_to_text, osl_backend
-                    )
-                    if osl is not None:
-                        # Same shape/key as the perf report output_sequence_lengths.
-                        entry["output_sequence_lengths"] = osl
-                        # Wall-clock of just this phase's tokenization (seconds);
-                        # summed across datasets for the accuracy report's total.
-                        entry["osl_tokenize_s"] = round(time.perf_counter() - t0, 3)
-            except Exception as e:  # noqa: BLE001 - optional blocks; never fail scoring
-                logger.warning(
-                    "Accuracy response counts/OSL skipped for %s: %s",
-                    eval_cfg.dataset_name,
-                    e,
-                )
-
-        accuracy_scores.append(entry)
-        logger.info(
-            f"Score for {eval_cfg.dataset_name}: {score} "
-            f"({n_repeats} repeats, complete={scorer_instance.complete})"
-        )
-
-    return accuracy_scores
-
-
-def finalize_benchmark(ctx: BenchmarkContext, bench: BenchmarkResult) -> None:
-    """Score accuracy, aggregate results, write JSON."""
-    result = bench.session
-    collector = bench.collector
-    report = bench.report
-
-    # Sibling profiling.json — kept separate so Report stays a pure
-    # snapshot-derived struct.
-    if bench.profiling is not None:
-        (ctx.report_dir / "profiling.json").write_text(
-            json.dumps(bench.profiling, indent=2)
-        )
-
-    # Write scoring artifacts + copy event log from tmpfs to disk (scorers read
-    # sample_idx_map.json + events.jsonl from here).
-    _write_scoring_artifacts(ctx, result, bench.tmpfs_dir)
-
-    # Accuracy scoring (one entry per accuracy dataset).
-    # Scoring runs before the report is written so the accuracy headline can be
-    # attached, but the report is written in the `finally` below so a scoring
-    # failure (e.g. lcb-service unreachable, missing eval subproject, bad extras)
-    # still leaves the perf run's result_summary.json / report.txt on disk
-    # instead of discarding them — then the exception propagates as before.
-    accuracy_scores: list[dict[str, Any]] = []
-    try:
-        accuracy_scores = _score_accuracy(ctx, result)
-    finally:
-        # Attach the per-dataset accuracy list so result_summary.json, the
-        # console summary, and report.txt all carry it (stays [] on a scoring
-        # failure).
-        if report is not None:
-            report = msgspec.structs.replace(report, accuracy=accuracy_scores)
-
-        # Display report if available (from MetricsAggregator pub/sub snapshot).
-        # result_summary.json is the self-complete machine-readable report
-        # (carries qps/tps/seeds/accuracy via Report.to_json); report.txt is the
-        # full human-readable dump; the console log shows the summary.
-        if report is not None:
-            report.display(fn=lambda s: logger.info(s), summary_only=True)
-            performance_dir = ctx.report_dir / "performance"
-            performance_dir.mkdir(parents=True, exist_ok=True)
-            report.to_json(save_to=performance_dir / "result_summary.json")
-
-            report_txt = ctx.report_dir / "report.txt"
-            with report_txt.open("w") as f:
-                report.display(fn=lambda s: print(s, file=f))
-                if bench.profiling is not None:
-                    _write_profiling_section(f, bench.profiling)
-            logger.info("Report written to %s", report_txt)
-
-    # Report metrics: prefer Report from MetricsSnapshot, fall back to SessionResult
+def _summarize_and_log_metrics(
+    ctx: BenchmarkContext,
+    report: Report | None,
+    result: SessionResult,
+    collector: ResponseCollector,
+) -> None:
+    """Log the run's headline metrics, preferring Report over SessionResult."""
     if report is not None and report.duration_ns is not None:
         perf_elapsed = report.duration_ns / 1e9
         total_issued = report.n_samples_issued
@@ -1695,48 +1058,48 @@ def finalize_benchmark(ctx: BenchmarkContext, bench: BenchmarkResult) -> None:
         if len(collector.errors) > 3:
             logger.debug(f"  ... +{len(collector.errors) - 3} more")
 
-    # Emit the accuracy results as a focused artifact under accuracy/. Perf
-    # rollups (qps/tps/latency percentiles) live in performance/result_summary.json
-    # and response/error text lives in events.jsonl, so neither is duplicated
-    # here. Written only when configured scoring produced entries.
-    if accuracy_scores:
-        # Plain cross-component mean of the per-dataset scores (3 datasets for
-        # gpt-oss, 1 for DeepSeek-R1); None when nothing numeric was scored.
-        avg_accuracy = average_accuracy(accuracy_scores)
-        # Total finalize-time spent tokenizing accuracy outputs for OSL (seconds),
-        # summed across datasets. Emitted whenever OSL was computed for at least
-        # one dataset — gating on the key's presence, not the rounded wall-clock,
-        # so a sub-millisecond total (tiny outputs) still records 0.0 rather than
-        # silently dropping the field.
-        osl_computed = any("osl_tokenize_s" in e for e in accuracy_scores)
-        osl_tokenization_s = round(
-            sum(e.get("osl_tokenize_s", 0.0) for e in accuracy_scores), 3
+
+def finalize_benchmark(ctx: BenchmarkContext, bench: BenchmarkResult) -> None:
+    """Score accuracy, aggregate results, write JSON."""
+    result = bench.session
+    collector = bench.collector
+    report = bench.report
+
+    # Sibling profiling.json — kept separate so Report stays a pure
+    # snapshot-derived struct.
+    if bench.profiling is not None:
+        (ctx.report_dir / "profiling.json").write_text(
+            json.dumps(bench.profiling, indent=2)
         )
-        accuracy_dir = ctx.report_dir / "accuracy"
-        accuracy_results_path = accuracy_dir / "accuracy_results.json"
-        accuracy_payload: dict[str, Any] = {}
-        if avg_accuracy is not None:
-            accuracy_payload["average_accuracy"] = avg_accuracy
-        if osl_computed:
-            accuracy_payload["osl_tokenization_s"] = osl_tokenization_s
-        accuracy_payload["accuracy_scores"] = accuracy_scores
-        # Atomic write so a crash mid-write can't leave truncated JSON the
-        # compliance checker would read as corrupt. Not swallowed: if scoring
-        # produced entries but they can't be persisted — the dir can't be made
-        # (OSError), the payload won't serialize (TypeError/ValueError, e.g. a
-        # numpy scalar left in a breakdown block), or the write fails — fail the
-        # run loudly rather than exit 0 with no accuracy artifact.
-        try:
-            accuracy_dir.mkdir(parents=True, exist_ok=True)
-            atomic_write_bytes(
-                accuracy_results_path,
-                json.dumps(accuracy_payload, indent=2).encode(),
-            )
-        except (OSError, TypeError, ValueError) as e:
-            raise ExecutionError(
-                f"Failed to write accuracy results to {accuracy_results_path}: {e}"
-            ) from e
-        logger.info(f"Saved: {accuracy_results_path}")
+
+    # Write scoring artifacts + copy event log from tmpfs to disk (scorers read
+    # sample_idx_map.json + events.jsonl from here).
+    _write_scoring_artifacts(ctx, result, bench.tmpfs_dir)
+
+    # Accuracy scoring (one entry per accuracy dataset). Scoring runs before the
+    # report is written so the accuracy headline can be attached, but the report
+    # is written in the `finally` below so a scoring failure (e.g. lcb-service
+    # unreachable, missing eval subproject, bad extras) still leaves the perf
+    # run's result_summary.json / report.txt on disk instead of discarding them —
+    # then the exception propagates as before.
+    accuracy_scores: list[dict[str, Any]] = []
+    try:
+        accuracy_scores = _score_accuracy(ctx, result)
+    finally:
+        # Attach the per-dataset accuracy list so result_summary.json, the
+        # console summary, and report.txt all carry it (stays [] on a scoring
+        # failure).
+        if report is not None:
+            report = msgspec.structs.replace(report, accuracy=accuracy_scores)
+        # Display the report + write result_summary.json / report.txt.
+        if report is not None:
+            _write_report_artifacts(ctx, report, bench.profiling)
+
+    _summarize_and_log_metrics(ctx, report, result, collector)
+
+    # Emit the accuracy results as a focused artifact under accuracy/. Written
+    # after the report artifacts so a write failure here can't discard them.
+    write_accuracy_results(ctx.report_dir, accuracy_scores)
 
 
 def run_benchmark(
